# Add direct calls to BLAS to compute SVDs