@fredrikekre 2024-05-31
I don't know how difficult it is to support both with and without `libblastrampoline`, but alternatively we could just wait with mergin this until there is a new Julia LTS version out (assuming @ronisbr 's immediate needs are fulfilled by https://github.com/ronisbr/StaticArraysBlasInterfaces.jl).