@mateuszbaran 2024-05-31
A very simple change to merge this soon would be to just conditionally load `blas.jl`:
```julia

@static if VERSION >= v"1.7"
    include("blas.jl")
end
```
and import `libblastrampoline` in that file. We don't have to make SVD calls faster on Julia 1.6.