@ronisbr 2024-06-02
Hi! Sorry, I missed the notifications!

Yes, I am totally fine to allow this improvement works only for Julia >= 1.7. I will make that change and submit a new commit.

Thanks!